v17.5.0c
The project source was converted from CommonJS to ESM, using Rollup to generate CommonJS bundles.

This is because CommonJS is best supported by the [version of Electron](https://github.com/AAAhs/electron-hevc) that we use in releases.

Using bundles, the app loads much fewer files, starting faster and reducing the size of the installers.

Includes bug fixes for the latest commit.